27 the gpio examples dont have a generic driver that could be reused for other pins

+++ drivers/inc/led2.h
@@ -1,12 +1,12 @@
 #ifndef LED2_H_
 #define LED2_H_
 
-#include "stm32f4xx.h"
+#include <stdint.h>
 
 void led2_init(void);
 void led2_on(void);
 void led2_off(void);
 void led2_toggle(void);
-int led2_get_state(void);
+uint8_t led2_get_state(void);
 
 #endif /* LED2_H_ */
